+++ ChangeLog.md
@@ -3,6 +3,7 @@
 * Feature #28: Add `ChangeLog.md` file.
 * Feature #27: Remove splash screen, people gonna hate it.
 * Feature #26: Newest, free items should be added to Explore as tabs.
+* Feature #30: Comment delete by comment poster.
 
 
 0.0.1 - 2015-05-23